+++ assets/css-generate/generate-css-editor-rtl.php
@@ -69,10 +69,12 @@
  *
  * No need to load the 'assets/css-generate/custom-styles.php' as the file was already loaded
  * once while generating main stylesheet.
- * Use `reykjavik_custom_styles( true )` to generate Visual Editor Customizer styles only.
+ * Also, setting the 'editor' scope to generate Visual Editor Customizer styles only.
  */
 
-	$output .= "\r\n\r\n\r\n/**\r\n * Customize styles\r\n */\r\n\r\n" . reykjavik_custom_styles( true );
+	if ( is_callable( 'Reykjavik_Library_Customize_Styles::custom_styles' ) ) {
+		$output .= "\r\n\r\n\r\n/**\r\n * Customize styles\r\n */\r\n\r\n" . Reykjavik_Library_Customize_Styles::custom_styles( '', 'editor' );
+	}