+++ assets/css-generate/generate-css-rtl.php
@@ -71,7 +71,9 @@
  * once while generating main stylesheet.
  */
 
-	$output .= "\r\n\r\n\r\n/**\r\n * Customize styles\r\n */\r\n\r\n" . reykjavik_custom_styles();
+	if ( is_callable( 'Reykjavik_Library_Customize_Styles::custom_styles' ) ) {
+		$output .= "\r\n\r\n\r\n/**\r\n * Customize styles\r\n */\r\n\r\n" . Reykjavik_Library_Customize_Styles::custom_styles();
+	}